Refactor match views to separate components
