+++ frontend/src/routes/discuss.tsx
@@ -1,14 +1,8 @@
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import { useQuestionCategories } from '@/hooks/useQuestions';
+import { ErrorView } from '@/components/discuss/views/ErrorView';
+import { IdleView } from '@/components/discuss/views/IdleView';
+import { MatchedView } from '@/components/discuss/views/MatchedView';
+import { TimeoutView } from '@/components/discuss/views/TimeoutView';
+import { WaitingView } from '@/components/discuss/views/WaitingView';
 import {
   MATCH_ERROR_STATUS,
   MATCH_FOUND_MESSAGE_TYPE,
@@ -21,249 +15,7 @@ import {
   MATCH_WAITING_STATUS,
 } from '@/lib/consts';
 import { getToken } from '@/lib/utils';
-import { ArrowRight, Loader2, Timer } from 'lucide-react';
-import React, { useEffect, useRef, useState } from 'react';
-import { Link } from 'react-router-dom';
-
-const difficultyLevels = ['Easy', 'Medium', 'Hard'];
-
-interface IdleViewProps {
-  onStartMatching: (topic: string, difficulty: string) => void;
-}
-
-const IdleView: React.FC<IdleViewProps> = ({ onStartMatching }) => {
-  const [topic, setTopic] = useState('');
-  const [difficulty, setDifficulty] = useState('');
-  const [showErrors, setShowErrors] = useState(false);
-
-  const { data: topics, isLoading } = useQuestionCategories();
-
-  const handleStartMatching = () => {
-    if (topic && difficulty) {
-      onStartMatching(topic, difficulty);
-      setShowErrors(false);
-    } else {
-      setShowErrors(true);
-    }
-  };
-
-  if (isLoading) {
-    return (
-      <div className='flex flex-col items-center justify-center h-full'>
-        <Loader2 className='w-8 h-8 animate-spin' />
-        <p className='mt-2'>Loading...</p>
-      </div>
-    );
-  }
-
-  return (
-    <Card className='w-full max-w-md mx-auto'>
-      <CardHeader>
-        <CardTitle>Find a Coding Partner</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <div className='space-y-4'>
-          <div>
-            <Select onValueChange={setTopic}>
-              <SelectTrigger>
-                <SelectValue placeholder='Select a topic' />
-              </SelectTrigger>
-              <SelectContent>
-                {topics &&
-                  topics.map((t) => (
-                    <SelectItem key={t} value={t}>
-                      {t}
-                    </SelectItem>
-                  ))}
-              </SelectContent>
-            </Select>
-            {showErrors && !topic && (
-              <p className='text-sm text-red-500 mt-1'>Please select a topic</p>
-            )}
-          </div>
-
-          <div>
-            <Select onValueChange={setDifficulty}>
-              <SelectTrigger>
-                <SelectValue placeholder='Select difficulty' />
-              </SelectTrigger>
-              <SelectContent>
-                {difficultyLevels.map((d) => (
-                  <SelectItem key={d} value={d}>
-                    {d}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-            {showErrors && !difficulty && (
-              <p className='text-sm text-red-500 mt-1'>
-                Please select a difficulty level
-              </p>
-            )}
-          </div>
-
-          <Button onClick={handleStartMatching} className='w-full'>
-            Start Matching
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-};
-
-interface WaitingViewProps {
-  queuePosition: number;
-  onCancel: () => void;
-}
-
-const WaitingView: React.FC<WaitingViewProps> = ({
-  queuePosition,
-  onCancel,
-}) => {
-  const [elapsedTime, setElapsedTime] = useState(0);
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setElapsedTime((prevTime) => prevTime + 1);
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, []);
-
-  const formatTime = (seconds: number) => {
-    const mins = Math.floor(seconds / 60);
-    const secs = seconds % 60;
-    return `${mins}:${secs < 10 ? '0' : ''}${secs}`;
-  };
-
-  return (
-    <Card className='w-full max-w-md mx-auto'>
-      <CardHeader>
-        <CardTitle>Finding a Match</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <div className='space-y-4'>
-          <div className='flex justify-center'>
-            <Loader2 className='h-12 w-12 animate-spin text-primary' />
-          </div>
-          <Alert>
-            <AlertTitle>Searching for a match</AlertTitle>
-            <AlertDescription>
-              Your position in queue: {queuePosition}
-            </AlertDescription>
-          </Alert>
-          <div className='flex items-center justify-center space-x-2 text-sm text-muted-foreground'>
-            <Timer className='h-4 w-4' />
-            <span>Elapsed time: {formatTime(elapsedTime)}</span>
-          </div>
-          <p className='text-center text-sm text-muted-foreground'>
-            Please wait while we find a coding partner for you. This may take up
-            to 30 seconds.
-          </p>
-          <Button onClick={onCancel} variant='outline' className='w-full'>
-            Cancel Matching
-          </Button>
-        </div>
-      </CardContent>
-    </Card>
-  );
-};
-
-interface MatchedViewProps {
-  roomId: string;
-  onNewMatch: () => void;
-}
-
-const MatchedView: React.FC<MatchedViewProps> = ({ roomId, onNewMatch }) => (
-  <Card className='w-full max-w-md mx-auto'>
-    <CardHeader>
-      <CardTitle>Match Found!</CardTitle>
-    </CardHeader>
-    <CardContent>
-      <div className='space-y-4'>
-        <Alert>
-          <AlertTitle>Success</AlertTitle>
-          <AlertDescription>
-            Your room ID is: <span className='font-bold'>{roomId}</span>
-          </AlertDescription>
-        </Alert>
-        <p className='text-center text-sm text-muted-foreground'>
-          You've been matched with a coding partner. Use the room ID to join
-          your collaborative session.
-        </p>
-        <div className='space-y-4'>
-          <Button asChild className='w-full group'>
-            <Link to={`/room/${roomId}`}>
-              Join Room
-              <ArrowRight className='w-4 h-4 ml-2 group-hover:translate-x-2 transition-transform' />
-            </Link>
-          </Button>
-          <Button onClick={onNewMatch} variant='outline' className='w-full'>
-            Find Another Match
-          </Button>
-        </div>
-      </div>
-    </CardContent>
-  </Card>
-);
-
-interface TimeoutViewProps {
-  onRetry: () => void;
-}
-
-const TimeoutView: React.FC<TimeoutViewProps> = ({ onRetry }) => (
-  <Card className='w-full max-w-md mx-auto'>
-    <CardHeader>
-      <CardTitle>Matching Timed Out</CardTitle>
-    </CardHeader>
-    <CardContent>
-      <div className='space-y-4'>
-        <Alert variant='destructive'>
-          <AlertTitle>No Match Found</AlertTitle>
-          <AlertDescription>
-            We couldn't find a match within the time limit.
-          </AlertDescription>
-        </Alert>
-        <p className='text-center text-sm text-muted-foreground'>
-          Don't worry! This can happen when there aren't many users online. Feel
-          free to try again.
-        </p>
-        <Button onClick={onRetry} className='w-full'>
-          Try Again
-        </Button>
-      </div>
-    </CardContent>
-  </Card>
-);
-
-interface ErrorViewProps {
-  onRetry: () => void;
-}
-
-const ErrorView: React.FC<ErrorViewProps> = ({ onRetry }) => (
-  <Card className='w-full max-w-md mx-auto'>
-    <CardHeader>
-      <CardTitle>Error Occurred</CardTitle>
-    </CardHeader>
-    <CardContent>
-      <div className='space-y-4'>
-        <Alert variant='destructive'>
-          <AlertTitle>Something went wrong</AlertTitle>
-          <AlertDescription>
-            An error occurred while trying to find a match.
-          </AlertDescription>
-        </Alert>
-        <p className='text-center text-sm text-muted-foreground'>
-          This could be due to a network issue or a problem with our servers.
-          Please try again later.
-        </p>
-        <Button onClick={onRetry} className='w-full'>
-          Retry
-        </Button>
-      </div>
-    </CardContent>
-  </Card>
-);
+import React, { useEffect, useRef } from 'react';
 
 export default function DiscussRoute() {
   const [matchStatus, setMatchStatus] = React.useState('idle');
